requirements.txt
mcp
langgraph
langchain[openai]
python-dotenv
langchain-mcp-adapters
rich